pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dodoai"
version = "0.1.4"
description = "DoDo - A stateful agentic framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "griffe>=0.40",
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pdm.scripts]
test = "pytest tests/ -v"
test-unit = "pytest tests/unit -v"
test-integ = "pytest tests/integration -v"
test-cov = "pytest tests/ -v --cov=src/dodo --cov-report=html"
format = "black src/ tests/"
lint = "ruff check src/ tests/"
lint-fix = "ruff check --fix src/ tests/"
